playlist: Avoid crash in gtk_list_store_set with starttime.
In 6029009e a column starttime/STARTTIME_COL got added to a treeview
of type gint64 in data/playlist.ui.

Unfortunately in 159e5ae4 a call to gtk_list_store_set was added which
gets called with a plain 0 in the variadic argument list.

It looks like this 0 gets just treated as 32-bit integer.
Therefore the terminating -1 from the argument list is no longer
detected and more items get accessed as were given as arguments,
which leads at least at armhf to the crash described in below Debian bug.

Bug-Debian: https://bugs.debian.org/986432

+++ src/totem-playlist.c
@@ -1788,7 +1788,7 @@ totem_playlist_steal_current_starttime (TotemPlaylist *playlist)
 	 * hence the "steal" in the API name */
 	gtk_list_store_set (GTK_LIST_STORE (playlist->model),
 			    &iter,
-			    STARTTIME_COL, 0,
+			    STARTTIME_COL, (gint64) 0,
 			    -1);
 
 	return starttime;
